updated application yaml and added config files
