+++ src/main/resources/application.yaml
@@ -121,6 +121,12 @@ exit-survey:
   applicant-link: https://www.smartsurvey.co.uk/s/CivilDamages_ExitSurvey_Claimant/
   respondent-link: https://www.smartsurvey.co.uk/s/CivilDamages_ExitSurvey_Defendant/
 
+camunda:
+  bpm:
+    client:
+      base-url: ${FEIGN_CLIENT_CONFIG_REMOTERUNTIMESERVICE_URL:}
+      async-response-timeout: 29000
+
 ---
 
 spring:
